career_cyber_ai/requirements.txt
streamlit>=1.28.0
endee>=0.1.0
openai>=1.0.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
torch>=2.0.0
